Revert "Merge pull request #23 from Kelompok-13-Alterra/new-home"
This reverts commit 9c476cf52755e610dde8d00d51c8dc5931003e20, reversing
changes made to 8d9b15138a8f9a5c73ce4ec78034a6ea2b098e3e.
