+++ lib/screens/profile/circle_avatar_home_screen.dart
@@ -38,7 +38,7 @@ class _CircleAvatarHomeScreenState extends State<CircleAvatarHomeScreen> {
           )
         ],
       ),
-      body: const Column(),
+      body: Column(),
     );
   }
 }
