+++ lib/screens/profile/profile_screen.dart
@@ -23,13 +23,6 @@ class _ProfileScreenState extends State<ProfileScreen> {
           "Profile",
           style: TextTitleProfile().text24,
         ),
-        leading: IconButton(
-          icon: const Icon(Icons.arrow_back),
-          color: kBlackColor,
-          onPressed: () {
-            Navigator.pop(context);
-          },
-        ),
         actions: const [
           Padding(
             padding: EdgeInsets.only(
@@ -44,84 +37,82 @@ class _ProfileScreenState extends State<ProfileScreen> {
           ),
         ],
       ),
-      body: SingleChildScrollView(
-        child: Column(
-          // mainAxisAlignment: MainAxisAlignment.start,
-          // crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Column(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: [
-                Padding(
-                  padding: PaddingCircleAvatarProfile().padding11,
-                  child: Container(
-                    width: 88,
-                    height: 88,
-                    decoration: const BoxDecoration(
-                      shape: BoxShape.circle,
-                      color: Color(0xFF97CBFF),
+      body: Column(
+        // mainAxisAlignment: MainAxisAlignment.start,
+        // crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Column(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            children: [
+              Padding(
+                padding: PaddingCircleAvatarProfile().padding11,
+                child: Container(
+                  width: 88,
+                  height: 88,
+                  decoration: const BoxDecoration(
+                    shape: BoxShape.circle,
+                    color: Color(0xFF97CBFF),
+                  ),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.only(
+                  bottom: 8,
+                ),
+                child: Row(
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Text(
+                      "Michael Abraham",
+                      style: TextNameOfUser().text18,
                     ),
-                  ),
+                    IconButton(
+                      splashRadius: 10,
+                      onPressed: () {},
+                      icon: const Icon(Icons.edit),
+                      color: const Color(0xFF000000),
+                    ),
+                  ],
                 ),
-                Padding(
-                  padding: const EdgeInsets.only(
-                    bottom: 8,
-                  ),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Text(
-                        "Michael Abraham",
-                        style: TextNameOfUser().text18,
-                      ),
-                      IconButton(
-                        splashRadius: 10,
-                        onPressed: () {},
-                        icon: const Icon(Icons.edit),
-                        color: const Color(0xFF000000),
-                      ),
-                    ],
-                  ),
+              ),
+              Padding(
+                padding: const EdgeInsets.only(
+                  bottom: 8,
                 ),
-                Padding(
-                  padding: const EdgeInsets.only(
-                    bottom: 8,
-                  ),
-                  child: Text(
-                    "michaelabraham@gmail.com",
-                    style: TextGmail().text19,
-                  ),
+                child: Text(
+                  "michaelabraham@gmail.com",
+                  style: TextGmail().text19,
                 ),
-                Padding(
-                  padding: const EdgeInsets.only(
-                    bottom: 8,
-                  ),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: [
-                      Image.asset(
-                        'assets/image/meeting_room.png',
-                        width: 16,
-                        height: 16,
-                      ),
-                      const SizedBox(
-                        width: 10,
-                      ),
-                      Text(
-                        'No Company',
-                        style: TextNoCompany().text20,
-                      )
-                    ],
-                  ),
+              ),
+              Padding(
+                padding: const EdgeInsets.only(
+                  bottom: 8,
                 ),
-              ],
-            ),
-            cardProfile(context),
-            profileWidget(context),
-          ],
-        ),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Image.asset(
+                      'assets/image/meeting_room.png',
+                      width: 16,
+                      height: 16,
+                    ),
+                    const SizedBox(
+                      width: 10,
+                    ),
+                    Text(
+                      'No Company',
+                      style: TextNoCompany().text20,
+                    )
+                  ],
+                ),
+              ),
+            ],
+          ),
+          cardProfile(context),
+          profileWidget(context),
+        ],
       ),
     );
   }
