+++ lib/widgets/login/form_login_widget.dart
@@ -34,8 +34,6 @@ class _FormLoginState extends State<FormLogin> {
           content: Text('Selamat ${_emailController.text} berhasil login'),
         ),
       );
-      Navigator.pushNamedAndRemoveUntil(
-          context, '/bottom-nav', (route) => false);
     }
   }
